Add Dockerfile and docker-compose file
This commit finishes the initial setup for the Docker deployment on
the local development environment. I have added the Dockerfile to
build the Django web application into a Docker image and a
docker-compose file to setup the PostgreSQL database image that will be
used as the database backend for this app. I also added a docker-entrypoint
script that will be run by the app container on startup to migrate the
database and run the app server.

I also updated settings.py to make sure the app can read the environment
variable file (.env) and connect with the Postgres database container.

With this commit, the app can be run simply by using docker compose. The
guide for specifying the .env file will be added later in the README file.