Finished the "Add memory" page/feature
The "Add memory" page shows a Google Map with Google Embed Map API, then
let the user search for an address using the Google Places API, or let
the user click on a spot in the map to automatically fetch the address of
that spot using the Google Geocoding API.

Warning: To use the Google Map API, I need to embed the API key on the html
script tag to import the libraries, so the API key will be exposed on
the client side. I tried to find ways to import the libraries without
exposing the key but haven't found one yet. The Google Map API documentations
only says to set restriction on the API keys. So make sure that the Google API key
used in this project is restricted to the project's web domain,
only allow the key for the Google Map Javascript API, Embed Map API,
Places API and Geocoding API. For good measures, also set the API quota
properly to prevent the user from overusing the API since there isn't
a cache for the API calls yet.